多启动一个线程用于监听gh端透过udp发送过来的数据

+++ src/Utils.py
@@ -3,6 +3,8 @@ import sys
 import logging
 import os
 import datetime
+# import ipdb
+import shutil
 
 def create_logger(logger_name, log_file):
     '''
@@ -13,7 +15,6 @@ def create_logger(logger_name, log_file):
     Use   logger = logging.getLogger(logger_name) to obtain logging all
     through out
     '''
-    # Todo需要添加清理日志的功能（保留日志信息 一周）
     logger = logging.getLogger(logger_name)
     # Remove the stdout handler
     logger_handlers = logger.handlers[:]
@@ -31,7 +32,7 @@ def create_logger(logger_name, log_file):
     terminal_h.set_name('stdout')
 
     formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
-    tool_formatter = logging.Formatter(' %(levelname)s - %(message)s')
+    tool_formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
     file_h.setFormatter(formatter)
     terminal_h.setFormatter(tool_formatter)
 
@@ -42,11 +43,40 @@ def create_logger(logger_name, log_file):
 def delExpireLog(path):
     list_path = os.listdir(path)
     flag = 0
+    del_f = ""
     if len(list_path) > 30:
-        for f in enumerate(list_path):
-            #ToDo 自动识别不符合要求的文件，并进行删除，防止文件奔溃
-            date = f.split('.')[0].split('_')[-1]
-            year,month,day = date.split('-')
+        
+        for f in list_path:
+            if os.path.isdir(os.path.join(path,f)):
+                # ipdb.set_trace()
+                try:
+                    shutil.rmtree(os.path.join(path,f))
+                    # os.remove(os.path.join(path,f))
+                except:
+                    pass
+                continue  
+            fname, suffix = f.split('.')
+            if suffix != 'log':
+                try:
+                    os.remove(os.path.join(path,f))
+                except:
+                    pass
+                continue
+            fname_split = fname.split('_')
+            if len(fname_split) != 2:
+                try:
+                    os.remove(os.path.join(path,f))
+                except:
+                    pass
+                continue
+            f_date = fname_split[-1].split('-')
+            if len(f_date) !=3:
+                try:
+                    os.remove(os.path.join(path,f))
+                except:
+                    pass
+                continue
+            year,month,day = f_date[0],f_date[1],f_date[2]
             try:
                 date_num = int(year) * 10000 + int(month) * 100 + int(day)
             except:
@@ -60,8 +90,9 @@ def delExpireLog(path):
                 if date_num < min_date:
                     min_date = date_num
                     del_f = f
-        del_filepath = os.path.join(path, del_f)
-        os.remove(del_filepath)
+        if del_f:
+            del_filepath = os.path.join(path, del_f)
+            os.remove(del_filepath)
     else:
         pass
 
@@ -71,16 +102,15 @@ def create_TempFolder(filename):
     """
     user_path = os.path.expanduser("~")
     rpFolder = os.path.join(user_path, "roboticplus")
-    if not rpFolder:
+    if not os.path.exists(rpFolder):
         os.mkdir(rpFolder)
     rpLogerFolder = os.path.join(rpFolder,"log")
     if not os.path.exists(rpLogerFolder):
         os.mkdir(rpLogerFolder)
     current_date = str(datetime.date.today())  # 2020-05-22
     filename = filename + '_' + current_date + '.log'
-    # Todo 仅保留一个的日志文件，其他可以删除
     delExpireLog(rpLogerFolder)
-    return os.path.join(rpLogerFolder, filename)
+    return rpFolder, os.path.join(rpLogerFolder, filename)
 
 def readText2Dic(fname):
     data_dict = {}
